Import can/util/vdom if not part of project
Moves startup logic to it's own module and handles the case where the
loader.main does not depend on jquery or can/util/vdom/vdom in some way.
In the future we won't need to load vdom with steal and this logic won't
be necessary.
