+++ lib/configure_loader.js
@@ -34,11 +34,7 @@ module.exports = function(loader){
 				loader.renderingLoader.baseURL = baseURL;
 			}
 
-			return loaderImport.call(loader, "can/util/vdom/vdom", {
-				name: name
-			}).then(function(){
-				return loaderImport.apply(loader, args);
-			});
+			return loaderImport.apply(loader, args);
 		}
 		return loaderImport.apply(this, arguments);
 	};
