+++ lib/index.js
@@ -8,6 +8,7 @@ var addCookies = require( "./cookies" );
 var traceBundles = require("./bundles/index");
 var util = require("util");
 var defaults = require("lodash.defaults");
+var stealStartup = require("./startup");
 
 var debug = require("./zones/debug");
 var timeout = require("can-zone/timeout");
@@ -45,28 +46,8 @@ module.exports = function(cfg, options){
 	configureLoader(loader);
 	var bundleHelpers = traceBundles(loader);
 
-	function getMainModule(main){
-		// startup returns an Array in dev
-		main = Array.isArray(main) ? main[0] : main;
-		return main.importPromise || Promise.resolve(main);
-	}
-
-	var startup = steal.startup().then(function(main){
-		if(!doneSsr.globalDocument && typeof document !== "undefined") {
-			doneSsr.globalDocument = document;
-		}
-
-		// If live-reload is enabled we need to get a new main each
-		// time a reload cycle is complete.
-		if(loader.has("live-reload")) {
-			var importOpts = {name: "@ssr"};
-			loader.import("live-reload", importOpts).then(function(reload){
-				reload(function(){
-					startup = loader.import(loader.main).then(getMainModule);
-				});
-			});
-		}
-		return getMainModule(main);
+	var startup = stealStartup(steal, function(mainPromise){
+		startup = mainPromise;
 	});
 
 	var SSRStream = function(requestOrUrl){
